gratitude/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
faiss
sentence-transformer
numpy